Refactor OnlineEditor's data file entries into new components
